+++ packages/aws-library/src/aws_library/ec2/__init__.py
@@ -1,0 +1,11 @@
+from ._client import SimcoreEC2API
+from ._errors import EC2AccessError, EC2NotConnectedError, EC2RuntimeError
+
+__all__: tuple[str, ...] = (
+    "SimcoreEC2API",
+    "EC2AccessError",
+    "EC2NotConnectedError",
+    "EC2RuntimeError",
+)
+
+# nopycln: file
