+++ packages/aws-library/src/aws_library/ec2/_errors.py
