+++ packages/aws-library/tests/test_ec2_models.py
@@ -4,7 +4,7 @@
 
 
 import pytest
-from aws_library.ec2.models import AWSTagKey, AWSTagValue, EC2InstanceData, Resources
+from aws_library.ec2._models import AWSTagKey, AWSTagValue, EC2InstanceData, Resources
 from faker import Faker
 from pydantic import ByteSize, ValidationError, parse_obj_as
